+++ src/components/layout/header/Header.tsx
@@ -1,32 +1,41 @@
 import React from "react";
 import type { ReactNode } from "react";
 
-import classnames from "classnames";
+import type { Div } from "../../native";
 
 import { Action } from "../../ui/action/Action";
 import { EqualActions } from "../../ui/action/EqualActions";
-import { Div } from "../../native";
 
-import styles from "./Header.module.scss";
+import { After, Before, Contents, StyledHeader } from "./Header.styled";
 
 interface Props {
+    /**
+     * Should the main content (children) be centered?
+     */
     center?: boolean;
+    /**
+     * Styling variant.
+     */
     variant?: "toolbar" | "colored";
     /**
      * This indicates just how the borders are drawn and how content is aligned, not the actual position on the screen.
      * To set up position on the screen, you need to properly style the parent element.
      */
     position?: "top" | "left" | "right" | "bottom"; // @TODO disallow left/right if not inside StickyHeader?
+    /**
+     * Content to be displayed before the main content.
+     */
     before?: ReactNode;
+    /**
+     * Content to be displayed after the main content.
+     */
     after?: ReactNode;
 }
 
 const Header: React.FC<React.ComponentProps<typeof Div> & Props> = (props) => {
     const {
-        center,
-        variant,
         position = "top",
-        className: _className, before: _before, after: _after,
+        before: _before, after: _after,
         children,
         ...rest
     } = props;
@@ -36,36 +45,30 @@ const Header: React.FC<React.ComponentProps<typeof Div> & Props> = (props) => {
         return c && typeof c === "object" && "type" in c && c.type === Action;
     });
 
-    const cls = classnames(styles.header, {
-        [styles["header--center"] as string]: center,
-        [styles["header--toolbar"] as string]: variant === "toolbar",
-        [styles["header--colored"] as string]: variant === "colored",
-    }, styles[`header--${position}`], props.className);
-
     let contents = children;
     if (justActions) {
         const mode = position === "top" || position === "bottom" ? "horizontal" : "vertical";
-        contents = <EqualActions className={styles.actions as string} mode={mode}>{contents}</EqualActions>;
+        contents = <EqualActions mode={mode}>{contents}</EqualActions>;
     }
 
     let before: ReactNode;
     if (props.before != null) {
-        before = <div className={styles.before}>{props.before}</div>;
+        before = <Before>{props.before}</Before>;
     }
 
     let after: ReactNode;
     if (props.after != null) {
-        after = <div className={styles.after}>{props.after}</div>;
+        after = <After>{props.after}</After>;
     }
 
     return (
-        <Div className={cls} {...rest}>
+        <StyledHeader {...rest} position={position} data-header-position={position}>
             {before}
-            <div className={styles.contents}>
+            <Contents>
                 {contents}
-            </div>
+            </Contents>
             {after}
-        </Div>
+        </StyledHeader>
     );
 };