+++ src/components/layout/header/HeaderIconAction.tsx
@@ -1,13 +1,9 @@
 import React from "react";
 import type { ReactNode } from "react";
 
-import classnames from "classnames";
-
 import type { ICON } from "../../icons/Icon";
 
-import { Icon } from "../../icons/Icon";
-
-import styles from "./HeaderIconAction.module.scss";
+import { A, Btn, StyledIcon } from "./HeaderIconAction.styled";
 
 interface LinkProps { // @TODO extract? - same on list item
     href: string;
@@ -15,21 +11,42 @@ interface LinkProps { // @TODO extract? - same on list item
 }
 
 interface Props {
+    /**
+     * Icon to display. Can be a string (for build in icons) or a ReactNode (any icon component).
+     */
     icon?: ICON | Exclude<ReactNode, string>;
+    /**
+     * Function to call when the icon is clicked.
+     */
     onClick?: () => void;
+    /**
+     * If given it will make the icon a native link.
+     */
     href?: string;
+    /**
+     * If given it will make the icon a custom link. `to` will be passed to given `Link` component as `href`.
+     */
     to?: string;
+    /**
+     * Custom link component to use. Use with `to` prop.
+     */
     Link?: React.ComponentClass<LinkProps> | React.FC<LinkProps>;
-    label?: ReactNode;
-    className?: string;
+    /**
+     * Additional class name to apply.
+     */
+    className?: string | undefined;
 }
 
+/**
+ * Use this component if you need a clickable icon that stylistically fits the header.
+ * It can be a simple link, a button or a custom link component.
+ */
 const HeaderIconAction: React.FC<Props> = (props) => {
-    const { icon, label, href, to, Link, className, ...restProps } = props;
+    const { icon, href, to, Link, className, ...restProps } = props;
 
     let content: ReactNode = icon;
     if (typeof icon === "string") {
-        content = <Icon className={styles.icon as string} name={icon as ICON} />;
+        content = <StyledIcon name={icon as ICON} />;
     }
 
     if (to) {
@@ -37,20 +54,17 @@ const HeaderIconAction: React.FC<Props> = (props) => {
             throw new TypeError("`to` prop given without `Link` component");
         }
 
-        const aCls = classnames(props.className, styles.a);
-        return <Link href={to} {...restProps}><a className={aCls}>{content}</a></Link>;
+        return <Link href={to} {...restProps}><A className={props.className}>{content}</A></Link>;
     }
 
     if (href) {
-        const aCls = classnames(props.className, styles.a);
-        return <a href={href} className={aCls} {...restProps}>{content}</a>;
+        return <A href={href} className={props.className} {...restProps}>{content}</A>;
     }
 
-    const btnCls = classnames(props.className, styles.btn);
     return (
-        <button className={btnCls} onClick={props.onClick}>
+        <Btn className={props.className} onClick={props.onClick}>
             {content}
-        </button>
+        </Btn>
     );
 };